// frontend/src/app/core/api/generated/model/current-user-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CurrentUserResponse { 
    id?: string;
    username?: string;
    email?: string;
    roles?: Set<CurrentUserResponse.RolesEnum>;
}
export namespace CurrentUserResponse {
    export const RolesEnum = {
        Admin: 'ADMIN',
        InventoryManager: 'INVENTORY_MANAGER',
        Sales: 'SALES'
    } as const;
    export type RolesEnum = typeof RolesEnum[keyof typeof RolesEnum];
}